@@ -1,16 +1,17 @@
-apply plugin: 'com.android.application'
-
-//Firebase
-apply plugin: 'com.google.gms.google-services'
+plugins {
+    id 'com.android.application'
+    id 'com.google.gms.google-services'
+    id 'com.diffplug.spotless' version '7.2.1'
+}
 
 android {
-    compileSdkVersion 30
-    buildToolsVersion "29.0.3"
+    namespace 'com.sefricam.sefricam3' // NUEVO y obligatorio con AGP 8+
+    compileSdk 35 // Android 14 o 15
 
     defaultConfig {
         applicationId "com.sefricam.sefricam3"
-        minSdkVersion 24
-        targetSdkVersion 30
+        minSdk 24
+        targetSdk 34
         versionCode 5
         versionName "3.3"
 
@@ -27,37 +28,51 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    testImplementation 'junit:junit:4.13.1'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    
+    // UI
+    implementation 'androidx.appcompat:appcompat:1.7.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
 
-    //Other Libraries
+    // Testing
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
+
+    // Responsive sizing
     implementation 'com.intuit.sdp:sdp-android:1.0.6'
     implementation 'com.intuit.ssp:ssp-android:1.0.6'
 
-    //Firebase
-    // Import the BoM for the Firebase platform
-    implementation platform('com.google.firebase:firebase-bom:25.12.0')
-
-    // Declare the dependency for the Realtime Database library
-    // When using the BoM, you don't specify versions in Firwwqaq2ebase library dependencies
+    // Firebase (BoM)
+    implementation platform('com.google.firebase:firebase-bom:33.16.0') // última BoM recomendada
     implementation 'com.google.firebase:firebase-database'
+    implementation 'com.google.firebase:firebase-analytics'
+    implementation 'com.google.firebase:firebase-firestore'
+    implementation 'com.google.firebase:firebase-auth'
 
-    implementation 'com.google.firebase:firebase-core:18.0.0'
-    implementation 'com.google.firebase:firebase-analytics:18.0.0' //Google Analytics
-    implementation 'com.google.firebase:firebase-firestore:22.0.0' //Google FireStore
-    implementation 'com.google.firebase:firebase-auth:20.0.1' //Firebase Auth
-
-    //Parse
-    // Don't forget to change the line below with the latest version of Parse SDK for Android
+    // Parse SDK
     implementation "com.github.parse-community.Parse-SDK-Android:parse:1.25.0"
 
+    // Loombok
+    compileOnly 'org.projectlombok:lombok:1.18.30'
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
+}
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
 }
 
 repositories {
+    google()
     mavenCentral()
-    jcenter()
     maven { url 'https://jitpack.io' }
+}
+
+spotless {
+    java {
+        googleJavaFormat()
+        leadingSpacesToTabs()
+        trimTrailingWhitespace()
+        endWithNewline()
+    }
 }
