@@ -3,6 +3,9 @@ package com.sefricam.sefricam3;
 import java.io.Serializable;
 import java.util.ArrayList;
 
+import lombok.Getter;
+
+@Getter
 public class MetodosCaptura implements Serializable {
     private final int NumeroParticipantes;
     private final int NumeroMallas;
@@ -125,203 +128,4 @@ public class MetodosCaptura implements Serializable {
         Observaciones = observaciones;
     }
 
-    public int getNumeroMallas() {
-        return NumeroMallas;
-    }
-
-    public double getLongitudRed() {
-        return LongitudRed;
-    }
-
-    public boolean isCoto() {
-        return Coto;
-    }
-
-    public ArrayList<Boolean> getControlAgentes() {
-        return ControlAgentes;
-    }
-
-    public int getReclamosCamachuelo() {
-        return ReclamosCamachuelo;
-    }
-
-    public int getCimbelesCamachuelo() {
-        return CimbelesCamachuelo;
-    }
-
-    public int getCapturasCamachueloM() {
-        return CapturasCamachueloM;
-    }
-
-    public int getCapturasCamachueloH() {
-        return CapturasCamachueloH;
-    }
-
-    public int getReclamosJilguero() {
-        return ReclamosJilguero;
-    }
-
-    public int getCimbelesJilguero() {
-        return CimbelesJilguero;
-    }
-
-    public int getCapturasJilgueroM() {
-        return CapturasJilgueroM;
-    }
-
-    public int getCapturasJilgueroH() {
-        return CapturasJilgueroH;
-    }
-
-    public int getReclamosLugano() {
-        return ReclamosLugano;
-    }
-
-    public int getCimbelesLugano() {
-        return CimbelesLugano;
-    }
-
-    public int getCapturasLuganoM() {
-        return CapturasLuganoM;
-    }
-
-    public int getCapturasLuganoH() {
-        return CapturasLuganoH;
-    }
-
-    public int getReclamosPardComun() {
-        return ReclamosPardComun;
-    }
-
-    public int getCimbelesPardComun() {
-        return CimbelesPardComun;
-    }
-
-    public int getCapturasPardComunM() {
-        return CapturasPardComunM;
-    }
-
-    public int getCapturasPardComunH() {
-        return CapturasPardComunH;
-    }
-
-    public int getReclamosPicogordo() {
-        return ReclamosPicogordo;
-    }
-
-    public int getCimbelesPicogordo() {
-        return CimbelesPicogordo;
-    }
-
-    public int getCapturasPicogordoM() {
-        return CapturasPicogordoM;
-    }
-
-    public int getCapturasPicogordoH() {
-        return CapturasPicogordoH;
-    }
-
-    public int getReclamosPinzonComun() {
-        return ReclamosPinzonComun;
-    }
-
-    public int getCimbelesPinzonComun() {
-        return CimbelesPinzonComun;
-    }
-
-    public int getCapturasPinzonComunM() {
-        return CapturasPinzonComunM;
-    }
-
-    public int getCapturasPinzonComunH() {
-        return CapturasPinzonComunH;
-    }
-
-    public int getReclamosPinzonReal() {
-        return ReclamosPinzonReal;
-    }
-
-    public int getCimbelesPinzonReal() {
-        return CimbelesPinzonReal;
-    }
-
-    public int getCapturasPinzonRealM() {
-        return CapturasPinzonRealM;
-    }
-
-    public int getCapturasPinzonRealH() {
-        return CapturasPinzonRealH;
-    }
-
-    public int getReclamosPiquituerto() {
-        return ReclamosPiquituerto;
-    }
-
-    public int getCimbelesPiquituerto() {
-        return CimbelesPiquituerto;
-    }
-
-    public int getCapturasPiquituertoM() {
-        return CapturasPiquituertoM;
-    }
-
-    public int getCapturasPiquituertoH() {
-        return CapturasPiquituertoH;
-    }
-
-    public int getReclamosVerdecillo() {
-        return ReclamosVerdecillo;
-    }
-
-    public int getCimbelesVerdecillo() {
-        return CimbelesVerdecillo;
-    }
-
-    public int getCapturasVerdecilloM() {
-        return CapturasVerdecilloM;
-    }
-
-    public int getCapturasVerdecilloH() {
-        return CapturasVerdecilloH;
-    }
-
-    public int getReclamosVerdComun() {
-        return ReclamosVerdComun;
-    }
-
-    public int getCimbelesVerdComun() {
-        return CimbelesVerdComun;
-    }
-
-    public int getCapturasVerdComunM() {
-        return CapturasVerdComunM;
-    }
-
-    public int getCapturasVerdComunH() {
-        return CapturasVerdComunH;
-    }
-
-    public int getReclamosVerdSerrano() {
-        return ReclamosVerdSerrano;
-    }
-
-    public int getCimbelesVerdSerrano() {
-        return CimbelesVerdSerrano;
-    }
-
-    public int getCapturasVerdSerranoM() {
-        return CapturasVerdSerranoM;
-    }
-
-    public int getCapturasVerdSerranoH() {
-        return CapturasVerdSerranoH;
-    }
-
-    public String getObservaciones() {
-        return Observaciones;
-    }
-
-    public int getNumeroParticipantes() {
-        return NumeroParticipantes;
-    }
 }
